Fix duplicate settings code for wxBitmapButton
When using object inheritance all settings templates of the base classes get applied as well without the possibility to override that. Because there is no way to detect the current class inside the template (and because its bad design anyway to let the base class template to act differently depending on the concrete derived class) supply a wxNullBitmap in the wxBitmapButton constructor and let the settings template set the bitmap instead. Because of the inheritance a markup label may be applied as well which doesn't have an effect because it doesn't get displayed, but to keep the preview in sync with the generator templates it also applies the markup label.