Store markup setting in XRC
The setting gets only stored for wxButton but not wxBitmapButton. It is not so nice that wxBitmapButton does have this property as well and the code generator even sets it but this cannot be easily disabled because of the inheritance. Luckily it doesn't seem to have any effect, wxBitmapButton never shows a label.