PHP: Pass down additional wxPanel constructor arguments.
Add optional arguments $id, $pos, $size, $style and $name to the
generated constructor for PHP wxPanel forms.  Provide the attributes
from the project as default values to these constructor arguments.
Pass the arguments down to the parent constructor.

Note that the #ifnotnull check for $window_name is now omitted, since
wxEmptyString is correctly passed when no window_name is set.