Fix reading and writing of the subclass element and all other elements that are of type PT_PARENT
The method used to extract the value of a composed wxPGProperty always returns an empty result, only GenerateComposedValue() does return a value. All locations testing a boolean of such a composed property now use the same approach, compare against the property name. Because of the rather strange storage format of a composed boolean and especially the difference to the format of a non-composed boolean comments about the storage format have been added. Parsing the default values of PT_PARENT child elements from a XML file now works properly with empty child tags.