Fix assertions because of lost event handlers
The smart trick of the Component base class to pop the event handler if it doesn't match the object does not work because certain wxWidgets elements push event handlers themselves on child objects so this condition cannot be used to determine if the Component itself has pushed one. Remove the cleanup code from the base class and implement it in the derived class instead, this one knows for sure if it has pushed an event handler or not.

NOTE: This introduces an API break, plugins are now required to clean up their pushed event handlers, the base class doesn't do it anymore.