Optimize and simplify wxCollapsiblePane code generation
Instead of assuming a certain layout of construction code templates and search-and-replace them to inject the call for GetPane() use the same approach like for wxStaticBoxSizer, let the #wxparent macro detect the wxCollapsiblePane and return the correct code fragment.

This reverts the code generator parts of ed8f84c36f6223e5156d9452ea6d7fcf92dd6330.