Code synchronization and cleanup
Apply the same logic and code style to the other bitmap cases and do some formatting and whitespace cleanup.

TODO: That const_cast doesn't look right there and might explain that e.g. changing child values of source type wxArtProvider doesn't work correctly. It is probably not a good idea to modify the property in that method but move that functionality to OnSetValue().