Perform subclass forward declaration even if header is empty
This restores the behavior to before commit 81a138b but with the extension that this is only done if the subclass should be forward declared.
This allows to forward declare local elements in the project file so they are not required to be defined before elements that depend on them.