Trim single-line code lines from left when not keeping indents
This was only done for multi-line code lines resulting into leading whitespace if the TemplateParser returned a value without newlines but leading whitespace. This happens depending on the structure of the template, consecutive newlines and whitespace get condensed into a single space which is prepended or appended to the result by different parser operations.