Reorder processing of project script files to make wxformbuilder the first one
The order of the processing determines the order of the projects in the solution. Visual Studio marks the first project as starting project. The new order turns wxFormBuilder (the utilities script doesn't define a project, only some functions) into the starting project instead of the TiCPP library, which can't be started at all.