Move wx_config before all other definitions inside the configuration block
This is actually a method call that does modify some global variables that the following definitions, which are method calls as well, do reference. Without that move the values of the global variables depend on the last call of that method which is depended on the structure of the current script file and the order of execution of all script files.