Switch to eine/setup-msys2 action
Recent GitHub Runners include a MSYS2 installation at a default location, the setup script
uses that location to collect dependencies. The original numworks/setup-msys2 action cannot
use that installation but created a symlink at that location to its own installation.
This is not possible anymore, the forked action allows the usage of the preinstalled MSYS2.

Update the symlink patches for the included binutils library.