Remove removal of HKCU keys
This doesn't work reliable because it is only executed for the user executing the installer
which might be an admin user instead of the current user. Anyway this can delete the keys
only for one user and not the possible other present users.

If it is necessary to remove or update these keys this have to be done by the application.