Select wxFormBuilder files more explicit instead of full wildcard
Full wildcard selection of the output directory includes a lot of unwanted
intermediate files or files from the development system. Previously this was
solved by specifying excludes but this gets tedious the more different environments
need to be supported.

Instead be more specific and select the desired files more explicit, this is also
not so easy because of the multiple plugins with their subdirectories, use a macro
for simpler processing of these.