Link wxFormBuilder against libz
This is required by libbfd. Also remove the condition for libintl and link always on gcc.
This obsoletes the second build file patching step, the first step is still required
because PreMake always places the linkoptions in front of the links and this results
into wrong position of the wxWidgets libraries in the linking step.