Add two-step creation for wxFrame
Add property two_step_creation to enable the creation of the required methods.

The general approach of two-step creation is that the settings code template
defines the body of the Create() method instead of the parameterized Constructor.
The cons_decl and cons_def code templates define the methods in the proper order,
the after_connectevents code template is used to add the return statement at end.

Implemented for C++ only!