Install runtime dependencies on Windows
This is supported for MSVC and MSYS compilers only. Plain MINGW compilers
are difficult to support because their compiler runtime directories are
difficult to determine in a system independent way.

On Linux it is expected that the dependencies are present on the target system,
on MacOS the application bundle already contains all non-system dependencies.