Add code path to load plugins by name only
Rely on the platform to resolve the plugin shared library. This uses
RPATH where available. This code path is enabled for the CMake build
system only because the other systems don't set the RPATH.

This is a generic implementation and only tested on Linux, platform
specific additions might be necessary.