Generate revision from git
The revision is generated during the build phase to ensure it is always
up-to-date and to detect a dirty working copy. It is not an error if no
revision can be determined, in that case a warning gets output and an
empty revision is generated.

Because the current code appends the revision directly to the version,
an utility method to return the part after an embedded version was added.