Fix generation of revision file
Remove the explicit dependency on the target that generates the file,
this will cause an error if the file doesn't exist during the configuration
phase if it wasn't built yet. Mark the file as BYPRODUCTS of the target
that produces it, this will ensure that the file will be built in the build
phase and won't cause an error if it doesn't exist in the
configuration phase yet.