Fetch tags after checkout
The checkout action does not fetch any tags resulting in `git describe`
failing. The recommended solution `git fetch --prune --unshallow --tags`
of https://github.com/actions/checkout/issues/206 has the drawback
that a full clone gets created defeating the optimization attempts of
the action.

After a lot of try-and-error it was discovered that
`git fetch --prune --depth=1 --tags` transfers much less data at the
cost that `git describe` can only find a tag if it is directly on the
currently checked out ref. This is sufficient to describe release refs only.