Parse project version from CMakeLists.txt for Flatpak
Export as environment variable and insert it into the Flatpak filename.
Include the separators so that this variable can simply be appended.
In case the version can't be detected an empty value is returned.
Don't use a trailing separator, currently no architecture gets appended
so this value is the last part of the filename.