Copy and track resources during build
Create custom commands together with targets to copy the resources
during the build into the binary tree and to update them when necessary.
Use an individual group for definitions, code templates and icons.

During installation install the definitions and code templates by name
but the complete directory of the icons. The rationale behind this is
that in the future generator languages can maybe be configured.
Also the definitions and code templates should reside in a single directory.
The icons however might use a directory structure in the future, installing
such a structure by simply copying the whole directory is easier than
copying individual files.

The current binary tree target directories are temporary until staged build
has been added.