Limit path normalization to absolute and dots
This removes the resolution of tilde expressions on linux and shortcuts
on windows. This was the previous default and should not have been in
use to produce identical results on all platforms.

This also removes the expansion of environment variables which might be
unexpected but might also break existing projects that rely on this feature.