Fix parsing of font weight
Project files store the binary value of symbolic font weights, due to an
ABI break in wxWidgets 3.1.2 these values have changed. Since the new
values are distinct from the old ones, this can be detected and both
types can be supported.

Support the old and new values during reading but always write the values
of the wxWidets version the application is linked to, this reduces required
code changes. However, this can lead to silent project file conversions
on save!

Closes #713