Use source type directory to access sources
Using the type git will clone from the specified path.
This won't work for a local checked out repository because the submodules
are indeed not relative to that repository but inside it.
Using an absolute URL for submodules would fetch the submodules a second time.

Since the Flatpak manifest is designed to be run from a checked out
repository this seems to be the better approach instead of checking out
the submodules next to the main repository. While this would work on the
GitHub CI, the flatpak.sh script would fail on a local machine without
further preparations.

The directory approach has the downside that the flatpak builder
directories are actually included, so far this didn't cause any trouble
though.