Disable MSYS2 package updates
The logs show that the Windows workflows on average have twice the
running time of the other workflows. We do not really need the newest
packages of everything, using CMake building is much more mature and stable
now so build errors happening because of a changing environment shouldn't
be common.