Enable autocrlf for git inside msys2 environment
This fixes the issue that working copies get detected as dirty if the
repository is not properly normalized or doesn't contain the proper
configuration inside .gitattributes.

This seems to be caused by the fact that the repository gets checked
out from a windows shell using git of the windows installation that has
autocrlf enabled. The git inside the msys2 environment does not have
autocrlf enabled by default.