Add support for building on all current Ubuntu releases
Add newer package versions as alternatives, add supported distributions
to the changelog.

Notice:
The default build dependency resolver of sbuild will discard all
alternatives but the first one, if this one cannot be satisfied by the
distribution, another resolver needs to be used. The aptitude resolver
seems to be sufficient, it can be used by specifying
--build-dep-resolver=aptitude for sbuild.