Fix wxWidgets::wxWidgets target creation for CMake 3.27
The find module of this CMake version does create this target itself but only in local scope.
For the SDK, this target needs to be in global scope, promote the target if it is not global already.
When creating the target itself, use an INTERFACE IMPORTED type like the find module.