Add XRC filter type Passthrough
This filter type passes values between XRC and XFB
unmodified as text. This is useful for properties
that use the exact same representation in both formats.

Current code usually uses the type Text and relies
on the fact that the properties don't use values
that require escaping. The type Passthrough never
uses escaping and makes semantically clear that
values get passed unmodified.