Reorder and move project properties
Use a semantically better ordering by using an order that is roughly
based on how important the property is and when it is used during
the code generation process.

This moves the properties `encoding` and `use_microsoft_bom` from the
C++ category to the common category because in contrast to the
documentation these properties are in fact used by all code generators
except the XRC generator.

Improve help messages and add missing ones.