Allow `nonvisual` childs for all sizer components
Sizers are an essential part of the layout and are used in many places,
this adds a lot of more customization points to inject code into the
generated code. All of these locations should also work correctly for
the wxTimer component. This change adds a lot of usability without
cluttering up the data type definitions by adding the `nonvisual`
element to e.g. all container types.