Fix Lua RootWxParent processing
The current code uses a wrong approach to determine the root wxWidgets
element which was probably inspired by the Python and PHP generators.
But in these generators not the root element itself is determined but
the syntax element to access it, for Lua however the name of the element
itself must be determined.

The current code works around its own flaws by inserting the root element
marker in `ValueToCode()` and replacing it later with the root element name
and using some guesswork in `GetCode()` to handle some special cases. This
guesswork does not work in certain cases.

Fix this by not inserting the root element marker in `ValueToCode()` anymore
and removing the guesswork in `GetCode()`. Remove also the handling of
the undocumented macro `#parentname` which is not required anymore since
the previous commit. Revert the integrated sizer macro to use `#wxparent`
again and add the root element marker replacement code.